+++ miniconf_derive/src/field.rs
@@ -1,20 +1,43 @@
-use darling::{uses_lifetimes, uses_type_params, util::Flag, FromField};
+use darling::{
+    usage::{IdentSet, Purpose, UsesTypeParams},
+    uses_lifetimes, uses_type_params,
+    util::Flag,
+    FromField, FromMeta,
+};
 use proc_macro2::{Span, TokenStream};
 use quote::quote_spanned;
-use syn::spanned::Spanned;
+use syn::{parse_quote, spanned::Spanned};
+
+#[derive(Clone, Copy, Debug, PartialEq)]
+pub enum TreeTrait {
+    Key,
+    Serialize,
+    Deserialize,
+    Any,
+}
+
+#[derive(Debug, FromMeta, PartialEq, Clone, Default)]
+struct Deny {
+    serialize: Option<String>,
+    deserialize: Option<String>,
+    ref_any: Option<String>,
+    mut_any: Option<String>,
+}
 
 #[derive(Debug, FromField, Clone)]
 #[darling(attributes(tree))]
 pub struct TreeField {
     pub ident: Option<syn::Ident>,
-    pub ty: syn::Type,
+    ty: syn::Type,
     pub skip: Flag,
-    pub typ: Option<syn::Type>,
-    pub validate: Option<syn::Expr>,
-    pub get: Option<syn::Expr>,
-    pub get_mut: Option<syn::Expr>,
-    pub rename: Option<syn::Ident>,
-    pub defer: Option<syn::Expr>,
+    typ: Option<syn::Type>,
+    validate: Option<syn::Expr>,
+    get: Option<syn::Expr>,
+    get_mut: Option<syn::Expr>,
+    rename: Option<syn::Ident>,
+    defer: Option<syn::Expr>,
+    #[darling(default)]
+    deny: Deny,
 }
 
 uses_type_params!(TreeField, ty, typ);
@@ -30,6 +53,35 @@ impl TreeField {
 
     pub fn typ(&self) -> &syn::Type {
         self.typ.as_ref().unwrap_or(&self.ty)
+    }
+
+    pub fn bound(&self, traite: TreeTrait, type_set: &IdentSet) -> Option<TokenStream> {
+        if self
+            .uses_type_params(&Purpose::BoundImpl.into(), type_set)
+            .is_empty()
+        {
+            None
+        } else {
+            let bound: Option<syn::TraitBound> = match traite {
+                TreeTrait::Key => Some(parse_quote!(::miniconf::TreeKey)),
+                TreeTrait::Serialize => self
+                    .deny
+                    .serialize
+                    .is_none()
+                    .then_some(parse_quote!(::miniconf::TreeSerialize)),
+                TreeTrait::Deserialize => self
+                    .deny
+                    .deserialize
+                    .is_none()
+                    .then_some(parse_quote!(::miniconf::TreeDeserialize<'de>)),
+                TreeTrait::Any => (self.deny.ref_any.is_none() || self.deny.mut_any.is_none())
+                    .then_some(parse_quote!(::miniconf::TreeAny)),
+            };
+            bound.map(|bound| {
+                let ty = self.typ();
+                quote_spanned!(self.span()=> #ty: #bound,)
+            })
+        }
     }
 
     pub fn name(&self) -> Option<&syn::Ident> {
@@ -99,43 +151,67 @@ impl TreeField {
 
     pub fn serialize_by_key(&self, i: Option<usize>) -> TokenStream {
         // Quote context is a match of the field index with `serialize_by_key()` args available.
-        let getter = self.getter(i);
-        quote_spanned! { self.span()=>
-            #getter
-                .and_then(|value|
-                    ::miniconf::TreeSerialize::serialize_by_key(value, keys, ser)
-                )
+        if let Some(s) = &self.deny.serialize {
+            quote_spanned! { self.span()=> ::core::result::Result::Err(
+                ::miniconf::Traversal::Access(0, #s).into())
+            }
+        } else {
+            let getter = self.getter(i);
+            quote_spanned! { self.span()=>
+                #getter
+                    .and_then(|value|
+                        ::miniconf::TreeSerialize::serialize_by_key(value, keys, ser)
+                    )
+            }
         }
     }
 
     pub fn deserialize_by_key(&self, i: Option<usize>) -> TokenStream {
         // Quote context is a match of the field index with `deserialize_by_key()` args available.
-        let getter_mut = self.getter_mut(i);
-        let validator = self.validator();
-        quote_spanned! { self.span()=>
-            #getter_mut
-                .and_then(|item|
-                    ::miniconf::TreeDeserialize::<'de>::deserialize_by_key(item, keys, de)
-                )
-                #validator
+        if let Some(s) = &self.deny.deserialize {
+            quote_spanned! { self.span()=> ::core::result::Result::Err(
+                ::miniconf::Traversal::Access(0, #s).into())
+            }
+        } else {
+            let getter_mut = self.getter_mut(i);
+            let validator = self.validator();
+            quote_spanned! { self.span()=>
+                #getter_mut
+                    .and_then(|item|
+                        ::miniconf::TreeDeserialize::<'de>::deserialize_by_key(item, keys, de)
+                    )
+                    #validator
+            }
         }
     }
 
     pub fn ref_any_by_key(&self, i: Option<usize>) -> TokenStream {
         // Quote context is a match of the field index with `get_mut_by_key()` args available.
-        let getter = self.getter(i);
-        quote_spanned! { self.span()=>
-            #getter
-                .and_then(|item| ::miniconf::TreeAny::ref_any_by_key(item, keys))
+        if let Some(s) = &self.deny.ref_any {
+            quote_spanned! { self.span()=> ::core::result::Result::Err(
+                ::miniconf::Traversal::Access(0, #s).into())
+            }
+        } else {
+            let getter = self.getter(i);
+            quote_spanned! { self.span()=>
+                #getter
+                    .and_then(|item| ::miniconf::TreeAny::ref_any_by_key(item, keys))
+            }
         }
     }
 
     pub fn mut_any_by_key(&self, i: Option<usize>) -> TokenStream {
         // Quote context is a match of the field index with `get_mut_by_key()` args available.
-        let getter_mut = self.getter_mut(i);
-        quote_spanned! { self.span()=>
-            #getter_mut
-                .and_then(|item| ::miniconf::TreeAny::mut_any_by_key(item, keys))
+        if let Some(s) = &self.deny.mut_any {
+            quote_spanned! { self.span()=> ::core::result::Result::Err(
+                ::miniconf::Traversal::Access(0, #s).into())
+            }
+        } else {
+            let getter_mut = self.getter_mut(i);
+            quote_spanned! { self.span()=>
+                #getter_mut
+                    .and_then(|item| ::miniconf::TreeAny::mut_any_by_key(item, keys))
+            }
         }
     }
 }